iommufd/fault: Do not hold lock around copy_from/to_user
Either of copy_from/to_user() may sleep if pagefaults are enabled. Thus,
either of them can take a long time to wait for the requested user pages
to swap in. Since both of them are protected by fault->mutex, that could
block the iommufd_fault_iopf_handler() that wants to take the mutex, and
then a shared IRQ handler in the IOMMU driver to back up, and eventually
a potential global DOS.

Provide two list manipulation helpers with a faster spinlock protection:
 - Read out the first item of the deliver list
 - Restore an item back to the head of the list

Note that the fault->response xarray has an internal lock, thus no extra
protection is required.

Fixes: 07838f7fd529 ("iommufd: Add iommufd fault object")
Cc: stable@vger.kernel.org
Suggested-by: Jason Gunthorpe <jgg@nvidia.com>
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>