cover-letter: iommufd: Allocate attach_handle for any HWPT
An attach_handle is a hwpt->fault specific thing, so the fault.c has been
holding a set of attach/detach/replace helpers for attach_handle and iopf
enabling/disabling.

Now, both the iommufd_msi series [1] and the attach_pasid series [2] want
to reuse the attach_handle for non-fault cases. Rework the attach_handle,
to prepare for either of the series. It should be merged with either that
gets merged first.

[1] https://lore.kernel.org/linux-iommu/cover.1736550979.git.nicolinc@nvidia.com/
[2] https://lore.kernel.org/linux-iommu/20241219132746.16193-1-yi.l.liu@intel.com/

Patches in this series are from [1] with some fixes addressing the review
comments.

Changelog
v1 (since [1])
 * Rebase on v6.14-rc1
 * Unwrap the fault attach/detach/replace helpers
 * Split the attach_handle patch to make it clearer

Thanks
Nicolin