iommufd: Abstract iopt_pin_pages and iopt_unpin_pages helpers
The new vQUEUE object will be added for HW to access the guest memory, for
HW-accelerated virtualization feature. It needs to ensure the guest memory
pages are pinned when HW accesses them and they are contiguous in physical
address space.

This is very like the existing iommufd_access_pin_pages() that outputs the
pinned page list for the caller to test its contiguity.

Move those code from iommufd_access_pin/unpin_pages() and related function
for a pair of iopt helpers that can be shared with the vQUEUE allocator.

Since the vQUEUE allocator will be a user-space triggered ioctl function,
WARN_ON would not be a good fit in the new iopt_unpin_pages(), thus change
them to use WARN_ON_ONCE instead.

Rename check_area_prot() to align with the existing iopt_area helpers, and
inline it to the header since iommufd_access_rw() still uses it.

Reviewed-by: Pranjal Shrivastava <praan@google.com>
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>