iommu/arm-smmu-v3-iommufd: Support implementation-defined hw_info
Repurpose the @__reserved field in the struct iommu_hw_info_arm_smmuv3,
to an HW implementation-defined field @impl.

This will be used by Tegra241 CMDQV implementation on top of a standard
ARM SMMUv3 IOMMU. The @impl will be only valid if @flags is set with an
implementation-defined flag.

Thus in the driver-level, add an hw_info impl op that will return such
a flag to use the impl field.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>