cover-letter: Fix iommufd selftest FAIL and warnings with v6.16
A few selftest harness changes being merged to v6.16, which exposed some
bugs and vulnerabilities in the iommufd selftest code. Fix them properly.

Note that the patch fixing the build warnings at mfd is not ideal, as it
has possibly hit some corner case in the gcc:
https://lore.kernel.org/all/aEi8DV+ReF3v3Rlf@nvidia.com/

This is on github:
https://github.com/nicolinc/iommufd/commits/iommufd_selftest_fixes-v6.16

Changelog:
v2
 * Add "Reviewed-by" from Jason
 * Only use kfree() in the teardown()
 * Add an mmap_buffer_size for readability
v1
 https://lore.kernel.org/all/cover.1750049883.git.nicolinc@nvidia.com/

Thanks
Nicolin